Moved the direcory chooser to swing helpers
